Remove Travis CI configuration
The testing part of this is rendered obsolete by GH Actions.

The deployment to PyPI is not currently handled by GH Actions, but I think
that's a lower priority as there haven't been any tags laid down in years. We
should probably revisit this mechanism when the next release is due.