Compute moments for thin islands differently
If fitted axes are LARGER than the clean beam axes, they can be deconvolved from it.

The rest of this commit involves considerable changes in measuring islands, triggered by a logger warning, from line 1117 in image.py when processing a TBB image: an infinite error in declination. This was caused by computing moments for 'thin' islands in the same way as for 'thick' islands. This yielded an extreme small semi-minor axis (~1e-8 in pixel coordinates). When such an unrealistically small semi-minor axis is propagated into the error analysis, the end of the declination error bar can be outside the image. In that case wcsp2s can give 'inf'.

To fix this one has to reconsider computing moments for thin islands: it does not make sense to compute these if the minimal width is 1, as in the above cases. Actually, a minimal width of 2 is also dubious, since deriving 6 Gaussian parameters from four pixel values does not really make sense. That is why we here impose the same constraints wrt minimal width as for Gaussian fitting: it should be at least 3 (> 2).

For all other cases, i.e. 'thin' islands, only derive the peak pixel value (times fudge_max_pix) and the barycenter position. The Gaussian shape parameters are copied from the clean beam. These computations can never crash.

We have decided to compute this upfront, i.e. in all cases, also for 'thick' islands. This causes a duplication in computing the peak pixel value and barycenter position, since these quantities are also computed in fitting.moments. However, an investigation of the same AARTFAAC TBB image showed that 519 of 600 sources were 'thin', when analysis threshold = detection threshold = 5 * rms noise. Therefore, the duplication of these computations seems limited.