Calculate signal-to-noise ratios of detections
This is for VECTORIZED=True. This was still missing from the vectorized part of PySE.
It seemed easiest to add this to 'insert_island_data', but requires the noise map as
an extra input argument. The dummy array can now actually be used when implemented as
float32 instead of int32. The significance (sig) of the detection is defined as the
maximum signal-to-noise ratio of the spectral brightness to the local noise, i.e. the
local standard deviation of the background pixel values, across all island pixels.
Added some clarification here and there.