First step in computing chisq and reduced_chisq vectorized
After some considerations it turns out that chisq and reduced_chisq are best calculated in 'fitting.moments_enhanced'. That, however, also requires the availability of 'noise_islands', i.e. the rms noise values at the positions of the sources, i.e. at the positions of the pixels comprising 'island_data', in that function. In turn that meant that we need some array with rms noise values similar to 'islands', which contains spectral brightnesses at the positions of the source. To fill 'noise_islands', 'insert_island_data' was expanded and the calculation of the significance of the detection was removed, since this is also best calculated in 'fitting.moments_enhanced'. In affectuating this change, it turns out that 'islands', 'island_data' and 'insert_island_data' are no longer accurate terms since an island can be an island of spectral brightnesses of a source or an island of rms noise values. Hence, the changes in variable and function names. The actual computations of sig, chisq and reduced_chisq will have to covered in upcoming commits. The current commit also includes some corrections and clarifications in comments. Some arrays were initialized later, i.e. closer to the function call. The 'noises' array has the same shape as the 'sources' array, which was formerly called the 'islands' array.