Generate ground truth data
for a new (accelerated) kappa*sigma clipper (to be implemented, so coming up). Our current stats.sigma_clip seems accurate, also in crowded fields, so we use it to generate ground truth data for a new clipper. We use the 'deconvolved.fits' image in combination with 'test_source_measurements.SourceParameters.setUp' to generate the ground truth data.