'image_206-215-t0002.fits' to generate ground
truth data for the upcoming accelerated kappa*sigma clipper. This is a good image to use, since it is an AARTFAAC TBB image, from Mattia Mancini (ASTRON). It was processed through 'python pyse.py  --detection=5 --analysis=5 --grid=128 --bmaj=0.208 --bmin=0.136 --bpa=15.619 --radius=1000 FITS_files/image_206-215-t0002-image.fits'. The output is a 2300² pixels image, which means that the subimage size of 128² pixels is not a divisor, which offers an extra test opportunity. Also, because 'radius=1000', part of the image will be masked, offering another test opportunity.