Added 'TestBackgroundCharacteristicsComplex' class
and renamed the 'TestBackgroundCharacteristics' class to 'TestBackgroundCharacteristicsSimple'. The tests within the 'TestBackgroundCharacteristicsComplex' class will probe the way clipping of masked data is handled since processing is done with 'radius=1000' on 'image_206-215-t0002.fits', an all-sky 2300² pixels AARTFAAC TBB image, i.e. this masks any pixel close to the horizon, or beyond the horizon, for that matter. Another test comes from the 128² pixels subimage-for-background-characteristics-determination size, since 128 is not a divisor of 2300, so the way 'leftover' pixels are handled is also tested. The mean background - usually estimated using its mode - is also tested. Wrt the 'TestBackgroundCharacteristicsSimple' class, some ground truth .npy files were renamed, moved to a separate folder and stored as .npz files which is more appropriate for masked data, since we also want to take account of masked grid values.