Clarification of 'find_true_std' and
debugging of stats.sigma_clip. 'data' does not seem to be a masked array while we would expect that if part of the imageis masked for a run with 'python pyse.py --detection=5 --analysis=5 --grid=128 --bmaj=0.208  --bmin=0.136 --bpa=15.619 --radius=1000 FITS_files/image_206-215-t0002-image.fits'. So we would expect a masked image because of '-radius=1000'. In 'image.__grids' we have 'useful_chunk = ndimage.find_objects(numpy.where(self.data.mask, 0, 1))' and ' useful_data = da.from_array(self.data[useful_chunk[0]].data, chunks=(self.back_size_x, y_dim))', which seems to assume that everything within the slice encompassing the masked pixels is unmasked, but this is not the case if the mask has the shape of a disc. This code needs to be corrected to be able to handle masked patches of sky that are not rectangular in shape or not aligned with the image pixel axes.