Adding 'np.ma.is_masked(chunk)' to the condition
(in an or statement) ensures that we get rms = 0 and mode = 0 for partially masked subimages. That makes sense since, if any pixel of a subimage is masked, one can no longer assume that the pixels used for κ×σ-clipping are centered on a grid node. Also, some reformatting, following linter recommendations.