Apply 'utils.nearest_nonzero' in
'ImageData.__grids' instead of masking entries in 'mode_grid' and 'rms_grid' since masks are not propagated in 'scipy.interpolate.interp1d', which meant that effectively zeroes were propagated which leads to an incorrect rmsmap, i.e. with values that are too low. 'mode_grid' and 'rms_grid' are now no longer MaskedArrays, but regular ndarrays.