The previous 'nearest_nonzero' implementation was
incorrect! It did not pass a cross-check with a similar function that uses 'scipy.spatial.KDTree'. This one does, at least for the non-ambiguous indices, i.e. for the indices of zeros in 'grid' that do not have an equal distance to multiple non-zeros. Note that we could also have used 'scipy.spatial.KDTree', but this one is 65 times faster for a 128² nodes grid. That may seem a lot, but the 'scipy.spatial.KDTree' version only takes about 7 ms, so in time the difference is negligible. Thanks @jhidding for explaining the problem with the previous version of 'nearest_nonzero'.