Add credit to ChatGPT 4.0
Although in this case it was merely for help in saving the cropped image in such a way that all header fields from the original file, i.e. 'GRB201006A_final_2min_srcs-t0002-image-pb.fits' were propagated adequately. Cropping a FITS image and saving it is pretty well documented: https://docs.astropy.org/en/stable/nddata/utils.html#saving-a-2d-cutout-to-a-fits-file-with-an-updated-wcs, but in our case 'hdu.data' from the original file was 4D, with two dimensions of length 1. When following the astropy instructions from the 'download_image_save_cutout' functionhandling the output file caused difficulties, coming from missing fields, not only for ds9, but also for PySE.