More robust version of nearest_nonzero
For the edge cases of grids of sizes 0 or 1. 'some_arr' seemed a more appropriate name than 'test_array'. Secondly, offer the option to replace elements in 'some_arr' based on the positions of zeros in 'rms' and the corresponding positions of nearest nonzeros in 'rms'. The elements of 'some_arr' at these indices will be used for replacement. We introduced this feature because a mean background can be zero (although highly unlikely), while a standard deviation (rms) cannot. Therefore the 'rms' array should provide directions for replacement.