Handle the edge case of a single element
rms background grid with that single value = 0. A standard deviation equal to zero means that no sensible way of thresholding based on noise statistics can be done and source detection should be cancelled. This is done my returning an empty grid, i.e. a grid of size 0. When ImageData._interpolate receives a size 0 grid as input, it will return a completely masked background map, such that no sources will be detected, as desired.