1) Compute the background grid on a suitable frame
within the frame that encompasses all unmasked data, i.e. within 'useful_chunk = ndimage.find_objects(np.where(self.data.mask, 0, 1))'. It is centered such that its offsets wrt the edges of 'useful_chunk' are almost equal along all sides. Also, it is as large as possible within 'useful_chunk' and an integer number of times 'back_size_x' along the row indices and an integer number of times 'back_size_y' along the column indices. The result of this is that a slightly smaller portion of the image data is used, but all subimages are now properly centred on background grid nodes. The dict returned by 'ImageData.__grids' now has an extra entry 'indices', which is used by 'ImageData._interpolate'.  The new 'nearest_nonzero' function is called and requires an extra argument, the rms grid. 2) Align within 80 characters per line. 3) We are looking for the mean of the background pixels, which we approximate by a mode estimation. So, strictly speaking, mean is a more appropriate term than mode. 4) 'ImageData._interpolate' should return a completely masked map if the grid has size 0. 5) In 'ImageData._interpolate' the unmasked tuple of slices encompassing the unmasked pixels does not have to be rederived, since it has been passed on as a call argument. 6) If there is no unmasked part of the rms map, we have to return the same number of arguments as in the opposite case.