Auxiliary function to make input data for
a guvectorized kappa * sigma clipper. The idea is that we have designed a function that will perform kappa * sigma clipping on a single flattened subimage, i.e. on a 1D ndarray. If we decorate that function with Numba's guvectorize decorator, we can use a 2D grid of subimages as input instead of a single flattened subimage. This means that the input to that guvectorized kappa * sigma clipper should be 3D. This function makes that 3D input, which is essentially a reshape using Numba with parallelization. “The 'make_subimages' function has been generated using ChatGPT 4.0. Its AI-output has been verified for correctness, accuracy and completeness, adapted where needed and approved by the author.”