Added all the machinery for a guvectorized
kappa * sigma clipper. It should mimick our current algorithm - 'sigma_clip' - as closely as possible, because we have established that it is quite accurate and its output has become part of the unit tests. Part of its accuracy comes from the correction of the standard deviation for clipping bias, i.e. because the standard deviation is calculated from a clipped distribution it will be biased low and we correct for that. This correction makes the kappa * sigma clipper converge twice as fast. The main difficulty in this coding effort is that the 'scipy.optimize.fsolve' root finder, needed to derive the clipping bias cannot be called from guvectorized code, i.e. Numba cannot compile it. That is why we asked ChatGPT 4.0 to come up with a root finder that can be compiled by Numba. 'f_sigma' is the same as 'find_true_std', it just has a different order of its arguments, so it can be removed in a next commit - and keep 'find_true_std' - if we simultaneously adapt the 'derivative' function and the order of the arguments in the call from within 'newton_1d_safeguard_sigma'. 'data_clipper_dynamic' needs a more complete docstring.