Added calls to the new guvectorized kappa * sigma
clipper, the previous algorithm has been commented out for now. Currently, we do not yet pass all unit tests, more specifically:
1) test/test_image.py::TestSimpleImageSourceFind::testSingleSourceExtraction
2) test/test_image.py::TestSimpleImageSourceFind::testWcsConversionConsistency
3) test/test_image.py::TestMaskedSource::testWholeSourceMasked
4) test/test_image.py::TestBackgroundCharacteristicsSimple::test_sigma_clip_deconvolved
5) test/test_image.py::TestBackgroundCharacteristicsComplex::test_sigma_clip_AARTFAAC_TBB_MASKED

The failed tests 1,2 and 3 occur at low noise levels, so most likely it is caused by rounding errors, which means that the background grid values are not computed accurately enough. This means we might need to adapt 'data_clipper_dynamic' to work with double precision instead of single precision floats. Possibly this can also fix 4 and 5, because the 'atol' and 'rtol' tolerances in 'numpy.allclose' are as small as 1e-8.