'f_sigma' and 'find_true_std' are equivalent
'f_sigma' and 'find_true_std' are equivalent. They just have different signs, but that does not matter for root finding. Let's keep 'find_true_std' and dump 'f_sigma', but that requires switching the second and third arguments if we want toretain the way this function is called as part of guvectorized kappa * sigma clipping.