We can now remove all 'sep' and 'dask' imports
We can now remove all 'sep' and 'dask' imports from the 'image' module as well as its use. Wrt 'sep' there was still the option for using it to determine the background grids, but the new guvectorized kappa * sigma clipper is just as fast and more accurate, so we removed the options to call 'sep' libraries. There was a 'hidden' call to 'sep' in 'self.background.globalrms' in 'ImageData.label_islands' that is now replaced by the average of the rms grid values, as determined by our new guvectorized kappa * sigma clipper. 'ImageData.compute_mode_and_rms_of_row_of_subimages' was called from a 'dask' array and is now no longer needed.