Attempt to replace 'scipy.interpolate.interp1d'
Attempt to replace 'scipy.interpolate.interp1d' by 'numpy.interp'. This code change gives an error: "ValueError: object too deep for desired array". Perhaps this is caused by 'grid', which is the second argument of 'scipy.interpolate.interp1d' and can be N-D, with a chosen 'axis' (keyword argument). For 'numpy.interp', 'grid' is the third argument and should be 1-D. Not sure how to fix this, since we want an interpolation across all columns of 'grid' for the desired equivalent of 'primary_interpolation' and across all columns of 'transposed' for the desired equivalent of 'perpendicular_interpolation'. Reverting to 'scipy.ndimage.zoom' may be an option.