Generate ground truth interpolated background data
Generate ground truth interpolated background data for "deconvolved.fits". This involves both the background means and standard deviations, interpolated from their grid values, with the grid nodes centered on the subimages.
Note that in order to generate the ground truth data, the new code with 'numpy.interp' has been removed. This routine is inconvenient, since it maps 1D ndarrays to 1D ndarrays, so it would have to be encapsulated in a loop.