'numpydoc' style docstrings for three functions
'numpydoc' style docstrings for three functions in the 'extract' module.
These are:
1) first_part_of_celestial_coordinates
2) insert_sources_and_noise
3) source_measurements_pixels_and_celestial_vectorised

For a number of parameters the data type was mentioned, but not the unit, e.g. pixels, degrees, arcseconds. For all parameters the units have been added, except for pixel values, since they do not differ from the observational image.