Unit test too strict
The "test_interpolation_AARTFAAC_TBB_MASKED" unit test turns out to be too strict: when running "image.ImageData._interpolate" with "kind='linear'" instead of "kind='slinear'" in "scipy.interpolate.interp1d" we should be getting the same results, since the algorithms seem to be the same algorithmically. However, at the level of the default "atol=1e-8" there appears to be a difference, probably due to rounding errors. Therefore we set "atol=1e-6".