There is no point in defining "xratio" and "yratio"
There is no point in defining "xratio" and "yratio", since they are equal to grid.shape[0] and grid.shape[1], respectively. In older versions of this code there were remainders in the sense that "self.back_size_x" and "self.back_size_y" were not divisors of "my_xdim" and "my_ydim", respectively, but this has been tackled recently in "image.ImageData.__grids".