Generalisation of the root finder
The idea is that 'stats.newton_raphson_root_finder' will not only be used for finding an unbiased standard deviation from a clipped distribution, but also to find a peak spectral brightness not obscured by pixellation effects. In implementing this generalisation, this root finder has to accommodate for functions with an arbitary number of arguments while the root has to be derived by changing only the first argument. That means, for 'stats.newton_raphson_root_finder', that we separate the initial guess and the upper and lower limits for that first argument from the other arguments, which are unpacked through "*args". Likewise, for a generalized derivative function, the first argument, "x" has been separated from the other arguments, again unpacked through "*args". However, the current implementation does not work and yields "TypeError: multiple values for argument 'tol'".