Extra argument for 'source_profile_and_errors'.
Extra argument for 'source_profile_and_errors' needed. This is the rms noise map within the rectangular area encompassing the island. We need this extra argument - 'rms' - to pass it on the 'fitting.goodness_of_fit', such that 'chisq' and 'reduced_chisq' can be computed slightly more accurately. Previously, only the noise level at the maximum pixel position was used.