'test_noisy_background' needs an extra argument
'test_noisy_background' needs an extra argument, which is the rms noise map at the chunk encompassing the island, as a consequence of the more accurate computation of 'chisq' and 'reduced_chisq' within 'fitting.goodness_of_fit'. For this unit test we had to construct this chunk of noise as 'rms=np.ones(self.mygauss.shape) * pixel_noise'. Also, replaced 'numpy.' by 'np.'.