Align flux density from moments with fits
When sources are fitted we compute the flux density from the peak spectral brightness and the fitted axes using the equivalent of equation (35) from the NVSS paper. Until now, we used the sum of all pixel values in the island when deriving an initial "guess" for the source parameters, but this too should be aligned with that equation (35).
To be clear, moments are not yet computed in this part of the code, this is only an initial "guess" for the source parameters, which will be retained when moments computations fail. Effectively, this will mean that the flux density will be equal to the peak spectral brightness unless "beamsize" is computed in a "deviant" manner.