"assert" statement redundant
As we know, the axes from moments - when computed in the simplest way - are biased low. This is an effect from the threshold that segments the island. We can correct for that, but that requires all pixel values within the island to be above the (local) analysis threshold. That is what the "assert" statement in "fitting.moments" from a previous commit tried to accomplish. However, you could consider this a redundant check, except for edge cases where the island is selected not by segmentation, but e.g. by selecting a patch of the sky around some set of celestial coordinates. Some unit tests select chunks of observational image data in that manner and we have to accept that "fitting.moments" will return biased Gaussian axes in those cases. Still, the sanity check "peak > threshold" is required or we will end up with math domain errors.