Removal NVSS bias correction -> known_result_fit
The removal of the bias correction from "extract.Paramset._condon_formulae" affects the ground truth peak spectral brightness, its error bar, chi-squared and reduced chi_squared for the "testSingleSourceExtraction" unit test. This commit adapts that ground truth data to our code.
Some background: we don't observe the "general fitting bias pulling the fitted peak in the direction of the local noise gradient" as reported in the NVSS paper, equation (34), at least not as a positive bias. We have therefore removed the second term on the right-hand side in the line of code that mimics equation (34), in a previous commit.

The new ground truth data shows both a lower chi-squared as well as a lower reduced chi-squared.