Ground truth data not derived in a robust way.
The ground truth data for the peak spectral brightness had been set at -0.00364 Jy/beam. It turned out this number came from negative bounds from a moments analysis; the "fitting.moments" code used to derive that number still used "data.mean()" in the conditional for deciding to derive a negative or a positive peak. Since "data.mean()" was negative, the bounds were negative.
In our updated code, we use "data[maxpos]" instead.
Moreover, we do not compute moments when one or more Gaussian parameters are fixed, consequently no bounds are set for the "testSourceAtGivenPosition_negative_spectral_brightness" unit test; the peak fit is unbounded, which yields this new ground truth peak spectral brightness of -0.00053898 Jy/beam. This includes the removal of the bias correction, i.e. the removal of the second term on the right-hand side of our implementation of equation (34) from the NVSS paper.