Accommodate for vectorized processing in unit test.
The "tests/test_image.py::TestSimpleImageSourceFind::testForceSourceShape" unit test will fail if we set "VECTORIZED = True" in the image module. Most likely, this has nothing to do with an actual error, but with a rounding error in the Numba compiled "measuring.moments_enhanced" enhanced function, coming from the use of single precision floats.