First step in making an artificial image
We want to expand our unit tests for vectorized source processing.
To do so, a large, e.g. 4K * 4K, with > 1e5 sources, would provide us with enough measurements for proper statistical tests.
Large images in a repo will quickly burn GH bandwidth, so we want to make artificial image, with sources inserted in the image plane on top of correlated noise, on the fly, as part of our GH CI.
To make correlated noise we need to convolve uncorrelated noise with a dirty beam (psf).
This one was produced by wsclean from a small MeasurementSet.