Make sure that bmaj, bmin and bpa are propagated.
This commit should enforce that if all three of bmaj, bmin and bpa are set as part of a ImgConf instance, they are propagated into "beam".
Previously, this propagation was not enforced.
"extract_metadata" will return metadata without beam information if not all three of bmaj, bmin and bpa are set as part of a ImgConf instance.