Make a new image on the fly for a new unit test.
The new unit test will test vectorized source processing for unresolved sources, i.e. with "force_beam=True".
In order to build proper statistics for the source measurements we make an artificial 4096² pixels image with 167_281 inserted sources on a "pseudo-regular" grid with a background of correlated noise.
"pseudo-regular", since we don't want all sources to be centred on the same position within a pixel, but randomly positioned within the area of a single pixel.
The correlated noise is made by convolving the psf with uncorrelated noise.
If we make an artificial image on the fly, we don't consume GH bandwidth.