Abort image processing if no beam.
It makes no sense to proceed with image processing if the beam has not been specified adequately, i.e. as a 3-tuple of floats (or doubles). The image module, being the "end point" for image processing - after the image has been loaded by other modules - seems the right place to abort image processing in that case.