Save ground truth as hdf5 file
An hdf5 file is slightly more convenient for loading into a Pandas df later, as part of the unit test.
We also want to load the measured source parameters from vectorized source processing in a Pandas df.
In this way, we can compare measurements and ground truth in a convenient way.