More elegant way to catch the case of zero sources
Previously, we had a line "elif num_islands > 0:" in "image.ImageData" to catch the case of zero detected sources.
We needed that conditional, since the line
"max_pixels = npixs.max()"
would give
"
ValueError: zero-size array to reduction operation maximum which has no identity
"
This is because "npixs" will become an array of size zero for zero detected sources.

This should be a slightly more elegant way of dealing with non-detections.