Fixed bug in "moments_enhanced"
1) We used to have:
"
if rounded_barycenter[0] in posx and rounded_barycenter[1] in posy:
    basepos = rounded_barycenter
    for i in (posx == rounded_barycenter[0]).nonzero()[0]:
        if posy[i] == rounded_barycenter[1]:
            basevalue = source_island[i]
            break
"
The problem here is that "posy[i] == rounded_barycenter[1]" may never be satisfied. Which means that "basevalue" is never set to "source_island[i]". Consequently, Numba set "basevalue=0", which led to a lot of errors.
The new code - ChatGPT generated - should fix that.

2) The other changes are formatting improvements from Black.