Provide functions for analytic convolution
When writing a regression test for vectorized source measurements, the sources added to the artificial images should match the ground truth peak brightnesses. That is not a problem if there are no subpixel offsets. It is also not a problem if we are inserting unresolved sources only as is the case for the "test_measured_vectorized_forced_beam" regression test.
It does become a problem for resolved sources when we want to mimick reality as closely as possible with subpixel offsets and we want to apply regular numerical convolution routines, such as "scipy.signal.convolve2d" or "scipy.signal.fftconvolve", to the actual - not pointlike - source. The reason is that we can then only control the true peak brightness of the convolved source at the pixel level, not at the subpixel level.
That is where "analytic convolution" comes in and ChatGPT 5.0 was helpful in providing these routines.
They should mimick the formula provided here:
https://math.stackexchange.com/questions/4422411/how-to-calculate-the-properties-of-the-resultant-2d-gaussian-from-the-deconvolut
but the formula provided by ChatGPT is much more concise because of the use of matrices.