"covariance_matrix" will be used outside tests
"covariance_matrix" was part of "testutil/convolve.py" since it was only used for making input images for regression testsw. However, we have seen that the calculation of error bars on deconvolved elliptical parameters has been inadequate. To make this calculation more accurate we need to transform covariance matrices for convolved elliptical parameters to its deconvolved counterparts using the following steps:
"
    meas_pa = 30
    meas = (6.0, 3.0, np.deg2rad(meas_pa))
    a, b, phi = meas
    beam_pa = -65
    beam = (3.0, 1.0, np.deg2rad(beam_pa))

    sigma_a = 0.05
    sigma_b = 0.04
    sigma_phi = np.deg2rad(0.5)

    C_p = np.diag([sigma_a**2, sigma_b**2, sigma_phi**2])

    C_S = cov_p_to_cov_S(C_p, meas[0], meas[1], meas[2])

    Sigma_meas = covariance_matrix(*meas)   # ensure same convention
    Sigma_beam = covariance_matrix(*beam)
    Sigma_dec  = Sigma_meas - Sigma_beam

    S_dec = np.array([Sigma_dec[0,0], Sigma_dec[1,1], Sigma_dec[0,1]])

    r_vec, C_r, ok = cov_S_to_cov_r(S_dec, C_S)

    sigma_a_dec = np.sqrt(C_r[0,0])
    sigma_b_dec = np.sqrt(C_r[1,1])
    sigma_phi_dec = np.sqrt(C_r[2,2])
"

So although we already have an algorithm for deconvolution - i.e. the "deconv" function - "covariance_matrix" will be convenient to perform deconvolution in terms of "Sigma" matrices, in order to derive errors bars on deconvolved elliptical parameters.