More accurate error bars on deconvolved parameters
Previously, the error bar on, say, the deconvolved semi-major axis, was estimated by adding the uncertainty of the convolved semi-major axis to its value and calculating the change in its deconvolved counterpart. Likewise by subtracting that uncertainty. The mean of those two changes was used as the uncertainty in the deconvolved semi-major axis.
This resulted in a standard deviation of the normalized residuals - from the true values - that was way too large, i.e. >> 1. This indicated that the error bars were not calculated properly, i.e. the uncertainties in the other two elliptical components should also be taken into account.
This commit attempts to do so, for the vectorized case, by calculating covariance matrices.
This is most conveniently done by using widths instead of half axes.
If not all three elliptical components can be deconvolved, the three error bars are set to NaNs.

Something similar will be needed for the non-vectorized case as well.