Expand proper error analysis to non-vectorized
Expand the proper error analysis for deconvolved elliptical components, that we had implemented for vectorized source measurements in commits d4e131f through 2823071, to non-vectorized source measurements.
What we see, from the "test_source_measurements.SourceParameters.testAllParameters" regression test, is that this leads to standard deviations of the normalized residuals of the deconvolved elliptical components very close to 1. This indicates that the error bars have been derived correctly. Previously, the standard deviations of the normalized residuals of the deconvolved axes were about 0.5, indicating that the error bars were too large. The new method had virtually no effect on the standard deviation of the normalized residuals of the deconvolved position angles.