Apply three different biases.
1) It makes sense to separate three different biases: "MAX_BIAS_BRIGHTNESSES=5.0", "MAX_BIAS_AXES = 10.0" and "MAX_BIAS_PAS = 1.0".
Previously, with the inaccurate calculation of the error bars on deconvolved elliptical parameters, we could set the maxximum biases for the "test_source_measurements.py::SourceParameters::testAllParameters" regression test as a single number, but with the recent improvement we found "signif_dev_smin=-7.411", i.e. larger, in absolute sense, than the old "MAX_BIAS=5.0".
Some background:
Peak brightnesses derived from Gaussian fits are biased low and the axes are biased high. The positions and major axes position angles are generally unbiased.
This is also true when these quantities are derived using "tweaked moments", but biases for peak brightnesses are smaller and the biases for the axes are larger. That is why we made these adjustments:
"
    MAX_BIAS_BRIGHTNESSES_SCALED /= 5.0
    MAX_BIAS_AXES_SCALED *= 1.5
    MAX_BIAS_PAS_SCALED *= 3.0
"
Deconvolved position angles may be biased, at least when they are derived using "tweaked moments".

2) We can increase the number of sources in the "test_measured_vectorized_free_shape" regression test to 90_000 without problems; no overlapping sources.