Source measurements should be performed vectorized
Source measurements should be performed vectorized by default, since it is much faster. Another upside is a lower bias than Gaussian fits on peak brightnesses - with the same sign as for Gaussian fits, i.e. biased low. A downside is a higher bias than Gaussian fits on the elliptical axes - with the same sign as for Gaussian fits, i.e. biased high.