Fixes "MaskedBackgroundBlind" failed unit test
With the default "vectorized = True" setting in the config module, running "test/test_image.py::TestMaskedBackground::testMaskedBackgroundBlind" will yield "TypeError: ufunc 'insert_sources_and_noise' not supported for the input types, and the inputs could not be safely coerced to any support". What is going on here is that if in "image.ImageData" "self.rmsmap" is completely masked, a few essential input values for vectorized source processing are not returned as Numpy ndarrays, but as lists or a Nones. That is corrected here.
Of course, this could have been handled differently, by not calling "extract.source_measurements_vectorised" at all when the background noise map is completely masked, since that precludes source extraction. This fix avoids adding a conditional.