"measuring.py" --> "measure.py"
Originally, this module was called "fitting.py". But it does more than Gaussian fits, it also computes (tweaked) moments. That is why it was renamed to "measuring.py". However, the other main modules use the simple present, as in "image.py" and "extract.py" instead of the present participle. That is why "measure.py" seems more appropriate.