Updated comments to reflect change in method
Previously, the "moments method" returned peak brightnesses as the maximum pixel value times a fudge factor to correct statistically for the fact that a peak will generally be positioned with some subpixel offset relative to the center of the pixel with the maximum value. Currently, we apply "tweaked moments" and that means that if a source comprises enough pixels, we will not use the maximum pixel method times a fudge factor. This has now been reflected in these two comments.