Include output of "pyse -h" in docs
We would like to have the output of "pyse -h" to be included in the documentation automatically, such that we do not have to change the documentation when the command-line interface changes. We decided to use a separate page "cli.rst" for this, since there is no page with command-line documentation yet.
ChatGPT provided directions on how to set this up:
1) Use the "sphinx-argparse" Python package: added as an optional dependency in "pyproject.toml".
2) Use the "sphinxarg.ext" extension.
3) Make a file "cli.rst" with these contents.