For "forced_beam": limit brightness correction
Fixes issue 180. What we have seen is that there are wider islands where the position of the pixel with the maximum value is several pixels away from the island's barycenter position. When the beam shape is forced to be as small as the clean beam, the extrapolation from the maximum pixel position to the barycenter position fails, i.e. the derived brightness value at the barycenter becomes huge. This fix limits the extrapolations from the rounded barycenter position to the barycenter position such that the subtended angle of interpolation remains at the subpixel level.
This bug did not show up when designing our two extensive "test_measured_vectorized_forced_beam" (167,281 inserted sources) and "test_measured_vectorized_free_shape" (90,000 inserted sources) regression tests, probably because these artificial sources are "well-formed" and do not share the peculiarities of sources in real aperture synthesis images.