"grid" does not only exist as cl parameter
"grid" does not only exist as command-line parameter, but also exists as an attribute of "ImgConf", i.e. it can be set by any user of the PySE libraries instead of the command line and that is the way it should be. This partially reverts commit a17bf64, i.e. the part described under 2) in the commit message of a17bf64.